Ghost race: ask it as a wizard step, not a sheet
Starting a workout asked two questions in sequence (activity, location)
and then hid a third — "am I racing a ghost?" — in a small card wedged
under the Indoor/Outdoor buttons, which opened a modal sheet with its own
nav stack, its own background gradient and its own card style. The race
read as an add-on to a screen about something else, and picking a target
left the flow entirely for ten seconds.

It's now two real steps in the same wizard:

  activity -> location -> normal vs ghost -> ghost options -> countdown

The picker moves out of GhostRaceSetupSheet into GhostRaceOptionsContent,
which carries no presentation chrome, so the wizard renders it inline
while the buddy lobby keeps showing it as a sheet — one picker, two
hosts, no drift. The lobby's call site is unchanged.

Details worth knowing:

- proceedPastRaceChoice() is the single door out of the race choice, so
  the once-ever live-presence interstitial can't be skipped by whichever
  step happens to come last.
- clearPreStartSteps() replaces the three hand-cleared booleans in the
  buddy hand-off and the recovery block, so a future step can't be
  forgotten in one of them.
- Asking about the ghost AFTER the activity is locked in also fixes an
  old oddity: going back and switching Run/Walk used to silently
  re-point an already-armed target at the other activity's stored time.
- The featured option card and the inline picker highlight by
  BRIGHTNESS, not hue. MADTheme.workoutColor("running") is the tracker
  gradient's own top stop, so a red-tinted CTA on that screen would
  vanish; the inline picker takes accent/accentForeground from its host
  and uses the same white-on-dark-red primary button the presence step
  already uses. The lobby sheet still gets the workout color.
- Steps share wizardTopBar/wizardHeader/wizardStep and slide
  directionally (forward from the trailing edge, back from the leading),
  so the flow reads as depth rather than four crossfades.

Not compile-verified: no Swift toolchain here, Xcode-only project.
